use signed ints

+++ baspacho/Solver.h
@@ -7,18 +7,18 @@
 
 struct Solver {
     Solver(CoalescedBlockMatrixSkel&& factorSkel,
-           std::vector<uint64_t>&& elimLumpRanges,
-           std::vector<uint64_t>&& permutation, OpsPtr&& ops);
+           std::vector<int64_t>&& elimLumpRanges,
+           std::vector<int64_t>&& permutation, OpsPtr&& ops);
 
     PermutedCoalescedAccessor accessor() const {
         return PermutedCoalescedAccessor(factorSkel.accessor(),
                                          permutation.data());
     }
 
-    void solveL(const double* matData, double* vecData, int stride,
+    void solveL(const double* matData, double* vecData, int64_t stride,
                 int nRHS) const;
 
-    void solveLt(const double* matData, double* vecData, int stride,
+    void solveLt(const double* matData, double* vecData, int64_t stride,
                  int nRHS) const;
 
     void factor(double* data, bool verbose = false) const;
@@ -29,26 +29,26 @@ struct Solver {
 
     void initElimination();
 
-    uint64_t boardElimTempSize(uint64_t lump, uint64_t boardIndexInSN) const;
+    int64_t boardElimTempSize(int64_t lump, int64_t boardIndexInSN) const;
 
     void factorLump(NumericCtx<double>& numCtx, double* data,
-                    uint64_t lump) const;
+                    int64_t lump) const;
 
     void eliminateBoard(NumericCtx<double>& numCtx, double* data,
-                        uint64_t ptr) const;
+                        int64_t ptr) const;
 
     void eliminateBoardBatch(NumericCtx<double>& numCtx, double* data,
-                             uint64_t ptr, uint64_t batchSize) const;
+                             int64_t ptr, int64_t batchSize) const;
 
     CoalescedBlockMatrixSkel factorSkel;
-    std::vector<uint64_t> elimLumpRanges;
-    std::vector<uint64_t> permutation;  // *on indices*: v'[p[i]] = v[i];
+    std::vector<int64_t> elimLumpRanges;
+    std::vector<int64_t> permutation;  // *on indices*: v'[p[i]] = v[i];
 
     OpsPtr ops;
     SymbolicCtxPtr symCtx;
     std::vector<SymElimCtxPtr> elimCtxs;
-    std::vector<uint64_t> startElimRowPtr;
-    uint64_t maxElimTempSize;
+    std::vector<int64_t> startElimRowPtr;
+    int64_t maxElimTempSize;
 };
 
 using SolverPtr = std::unique_ptr<Solver>;
@@ -59,11 +59,11 @@ struct Settings {
 };
 
 SolverPtr createSolver(const Settings& settings,
-                       const std::vector<uint64_t>& paramSize,
+                       const std::vector<int64_t>& paramSize,
                        const SparseStructure& ss, bool verbose = false);
 
 SolverPtr createSolverSchur(const Settings& settings,
-                            const std::vector<uint64_t>& paramSize,
+                            const std::vector<int64_t>& paramSize,
                             const SparseStructure& ss,
-                            const std::vector<uint64_t>& elimLumpRanges,
+                            const std::vector<int64_t>& elimLumpRanges,
                             bool verbose = false);
